feat(install): Configure Git using symbolic links
This commit introduces a `git_config` target to the Makefile that configures Git by creating symbolic links to the dotfiles in the repository.
This allows for easy management and updating of Git configuration across different machines.

The `git_config` target creates symbolic links for:
- .gitconfig
- .gitignore_global
- .gitconfig-osx (on macOS)
- .gitconfig-linux (on Linux)

The links are created in the user's home directory `$HOME`, ensuring that git uses the configurations defined in the dotfiles repository.